[Experimental] Attempt to improve syncing
The theory is that due to small precision issues the character isGrounded state might not be synced perfectly between client and server. By sending a jump command which ignores the isGrounded state this desync should be less noticable in theory. In practice though I haven't noticed much improvement. Let's give it a go anyways though